requirements.txt
annoy==1.15.1
beautifulsoup4==4.7.1
boto==2.49.0
boto3==1.9.129
botocore==1.12.129
bz2file==0.98
certifi==2019.3.9
chardet==3.0.4
Django==2.1.5
django-cors-headers==2.4.0
djangorestframework==3.9.0
docutils==0.14
gensim==3.7.1
geographiclib==1.49
geopy==1.19.0
gevent==1.4.0
greenlet==0.4.15
gunicorn==19.9.0
idna==2.8
jmespath==0.9.4
numpy==1.16.0
pandas==0.24.2
pymongo==3.7.2
python-crfsuite==0.9.6
python-dateutil==2.8.0
pytz==2018.9
pyvi==0.0.9.3
requests==2.21.0
s3transfer==0.2.0
scikit-learn==0.20.3
scipy==1.2.1
six==1.12.0
sklearn-crfsuite==0.3.6
smart-open==1.8.0
soupsieve==1.9
tabulate==0.8.3
tqdm==4.31.1
urllib3==1.24.1
